merge

+++ src/main/java/mx/edu/um/mateo/inscripciones/dao/PaqueteDao.java
@@ -24,7 +24,7 @@ public interface PaqueteDao {
      * @param id the paquete's id
      * @return paquete populated paquete object
      */
-    public Paquete obtiene(final Integer id);
+    public Paquete obtiene(final Long id);
     
     /**
      * Saves a paquete's information
@@ -36,6 +36,6 @@ public interface PaqueteDao {
      * Removes a paquete from the database by id
      * @param id the paquete's id
      */
-    public String elimina(final Integer id);
+    public String elimina(final Long id);
     
 }
